feat(injectify.autofac): move common implementation from Injectify.Autofac to Injectify common package
