+++ packages/Injectify.Autofac/Injectify.Autofac/Extensions/ServiceProviderExtensions.cs
@@ -15,15 +15,8 @@ namespace Injectify.Autofac.Extensions
         /// <param name="provider"></param>
         /// <param name="propInfo"></param>
         /// <returns></returns>
-        public static object GetByPropertyInfo(this IContainer provider, PropertyInfo propInfo)
-        {
-            if (propInfo.PropertyType?.GenericTypeArguments?.Any() ?? false)
-            {
-                //return provider.GetServices(propInfo.PropertyType?.GenericTypeArguments?.FirstOrDefault());
-            }
-
-            return provider.Resolve(propInfo.PropertyType);
-        }
+        public static object GetByPropertyInfo(this IContainer provider, PropertyInfo propInfo) =>
+            provider.Resolve(propInfo.PropertyType);
 
         /// <summary>
         /// 
@@ -31,14 +24,7 @@ namespace Injectify.Autofac.Extensions
         /// <param name="provider"></param>
         /// <param name="parameterInfo"></param>
         /// <returns></returns>
-        public static object GetByParameterInfo(this IContainer provider, ParameterInfo parameterInfo)
-        {
-            if (parameterInfo.ParameterType?.GenericTypeArguments?.Any() ?? false)
-            {
-                //return provider.GetServices(parameterInfo.ParameterType?.GenericTypeArguments?.FirstOrDefault());
-            }
-
-            return provider.Resolve(parameterInfo.ParameterType);
-        }
+        public static object GetByParameterInfo(this IContainer provider, ParameterInfo parameterInfo) =>
+            provider.Resolve(parameterInfo.ParameterType);
     }
 }
